+++ src/Ingredient/Attributes/Transform.php
@@ -29,6 +29,7 @@ use Attribute;
 use RuntimeException;
 
 use function class_exists;
+use function is_callable;
 
 /**
  * Class to implement attribute `Transform` to transform an ingredient before to put it into the bowl
@@ -44,16 +45,29 @@ use function class_exists;
 #[Attribute(Attribute::TARGET_PARAMETER)]
 class Transform
 {
+    /**
+     * @var callable|null
+     */
+    private $transformer = null;
+
     public function __construct(
         private ?string $className = null,
+        ?callable $transformer = null,
     ) {
         if (!empty($this->className) && !class_exists($this->className)) {
             throw new RuntimeException("Error the required class {$this->className} does not exist");
         }
+
+        $this->transformer = $transformer;
     }
 
     public function getClassName(): ?string
     {
         return $this->className;
+    }
+
+    public function getTransformer(): ?callable
+    {
+        return $this->transformer;
     }
 }